securing server routes

+++ backend/app/Controllers/AuthorsController.php
@@ -5,7 +5,7 @@ namespace App\Controllers;
 use Config\Database;
 use App\Models\Authors;
 
-class AuthorsController
+class AuthorsController extends Controller
 {
     public function readAll()
     {
@@ -39,7 +39,7 @@ class AuthorsController
         }
     }
 
-    public function readbyId()
+    public function readById()
     {
         // Méthode autorisée
         header("Access-Control-Allow-Methods: GET");
@@ -80,40 +80,44 @@ class AuthorsController
     {
         // Méthode autorisée
         header("Access-Control-Allow-Methods: POST");
+        if ($this->Authorization() == "[ROLE_ADMIN]") {
+            if ($_SERVER["REQUEST_METHOD"] === "POST") {
 
-        if ($_SERVER["REQUEST_METHOD"] === "POST") {
+                // On instancie la base de données
+                $database = new Database();
+                $db = $database->getConnexion();
 
-            // On instancie la base de données
-            $database = new Database();
-            $db = $database->getConnexion();
+                // On instancie l'objet Authors
+                $author = new Authors($db);
 
-            // On instancie l'objet Authors
-            $author = new Authors($db);
+                // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
+                $data = json_decode(file_get_contents("php://input"));
 
-            // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
-            $data = json_decode(file_get_contents("php://input"));
+                if (!empty($data->name)) {
+                    // On hydrate l'objet author
+                    $author->setName($data->name);
+                    $author->setNationality($data->nationality);
+                    $author->setBiography($data->biography);
 
-            if (!empty($data->name)) {
-                // On hydrate l'objet author
-                $author->setName($data->name);
-                $author->setNationality($data->nationality);
-                $author->setBiography($data->biography);
-
-                $result = $author->create();
-                if ($result) {
-                    http_response_code(201);
-                    echo json_encode(["result" => "Ok", "message" => "L'auteur' a été ajouté avec succès"]);
+                    $result = $author->create();
+                    if ($result) {
+                        http_response_code(201);
+                        echo json_encode(["result" => "Ok", "message" => "L'auteur' a été ajouté avec succès"]);
+                    } else {
+                        http_response_code(503);
+                        echo json_encode(["message" => "L'ajout de l'auteur' a échoué"]);
+                    }
                 } else {
-                    http_response_code(503);
-                    echo json_encode(["message" => "L'ajout de l'auteur' a échoué"]);
+                    //http_response_code(503);
+                    echo json_encode(["message" => "Les données ne sont pas complètes"]);
                 }
             } else {
-                //http_response_code(503);
-                echo json_encode(["message" => "Les données ne sont pas complètes"]);
+                http_response_code(405);
+                echo json_encode(["message" => "La méthode n'est pas autorisée"]);
             }
         } else {
-            http_response_code(405);
-            echo json_encode(["message" => "La méthode n'est pas autorisée"]);
+            http_response_code(401);
+            echo json_encode(["result" => "ERROR", "message" => "Autorisation refusée"]);
         }
     }
 
@@ -121,38 +125,42 @@ class AuthorsController
     {
         // Méthode autorisée
         header("Access-Control-Allow-Methods: PUT");
+        if ($this->Authorization() == "[ROLE_ADMIN]") {
+            if ($_SERVER["REQUEST_METHOD"] === "PUT") {
+                // On instancie la base de données
+                $database = new Database();
+                $db = $database->getConnexion();
+                // On instancie l'objet Authors
+                $author = new Authors($db);
 
-        if ($_SERVER["REQUEST_METHOD"] === "PUT") {
-            // On instancie la base de données
-            $database = new Database();
-            $db = $database->getConnexion();
-            // On instancie l'objet Authors
-            $author = new Authors($db);
+                // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
+                $data = json_decode(file_get_contents("php://input"));
 
-            // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
-            $data = json_decode(file_get_contents("php://input"));
+                if (!empty($data->id)) {
+                    $author->setId($data->id);
+                    $author->setName($data->name);
+                    $author->setNationality($data->nationality);
+                    $author->setBiography($data->biography);
 
-            if (!empty($data->id)) {
-                $author->setId($data->id);
-                $author->setName($data->name);
-                $author->setNationality($data->nationality);
-                $author->setBiography($data->biography);
-
-                $result = $author->update();
-                if ($result) {
-                    http_response_code(201);
-                    echo json_encode(["result" => "Ok", "message" => "L'auteur a été modifié avec succès"]);
+                    $result = $author->update();
+                    if ($result) {
+                        http_response_code(201);
+                        echo json_encode(["result" => "Ok", "message" => "L'auteur a été modifié avec succès"]);
+                    } else {
+                        http_response_code(503);
+                        echo json_encode(["message" => "La modification de l'auteur a échoué"]);
+                    }
                 } else {
-                    http_response_code(503);
-                    echo json_encode(["message" => "La modification de l'auteur a échoué"]);
+                    //http_response_code(503);
+                    echo json_encode(["message" => "Les données ne sont pas complètes"]);
                 }
             } else {
-                //http_response_code(503);
-                echo json_encode(["message" => "Les données ne sont pas complètes"]);
+                http_response_code(405);
+                echo json_encode(["message" => "La méthode n'est pas autorisée"]);
             }
         } else {
-            http_response_code(405);
-            echo json_encode(["message" => "La méthode n'est pas autorisée"]);
+            http_response_code(401);
+            echo json_encode(["result" => "ERROR", "message" => "Autorisation refusée"]);
         }
     }
 
@@ -160,35 +168,39 @@ class AuthorsController
     {
         // Méthode autorisée
         header("Access-Control-Allow-Methods: DELETE");
+        if ($this->Authorization() == "[ROLE_ADMIN]") {
+            if ($_SERVER["REQUEST_METHOD"] === "DELETE") {
+                // On instancie la base de données
+                $database = new Database();
+                $db = $database->getConnexion();
 
-        if ($_SERVER["REQUEST_METHOD"] === "DELETE") {
-            // On instancie la base de données
-            $database = new Database();
-            $db = $database->getConnexion();
+                // On instancie l'objet AUthors
+                $author = new AUthors($db);
 
-            // On instancie l'objet AUthors
-            $author = new AUthors($db);
+                // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
+                $data = json_decode(file_get_contents("php://input"));
 
-            // On récupère les informations envoyées et je décode le JSON pour que php puisse le lire
-            $data = json_decode(file_get_contents("php://input"));
+                if (!empty($data)) {
+                    $author->setId($data);
 
-            if (!empty($data)) {
-                $author->setId($data);
-
-                $result = $author->delete();
-                if ($result) {
-                    http_response_code(200);
-                    echo json_encode(["result" => "Ok", "message" => "La suppression de l'auteur a été effectué avec succès"]);
+                    $result = $author->delete();
+                    if ($result) {
+                        http_response_code(200);
+                        echo json_encode(["result" => "Ok", "message" => "La suppression de l'auteur a été effectué avec succès"]);
+                    } else {
+                        http_response_code(503);
+                        echo json_encode(["message" => "La suppression de l'auteur a échoué"]);
+                    }
                 } else {
-                    http_response_code(503);
-                    echo json_encode(["message" => "La suppression de l'auteur a échoué"]);
+                    echo json_encode(["message" => "Vous devez précisé l'identifiant de l'auteur"]);
                 }
             } else {
-                echo json_encode(["message" => "Vous devez précisé l'identifiant de l'auteur"]);
+                http_response_code(405);
+                echo json_encode(["message" => "La méthode n'est pas autorisée"]);
             }
         } else {
-            http_response_code(405);
-            echo json_encode(["message" => "La méthode n'est pas autorisée"]);
+            http_response_code(401);
+            echo json_encode(["result" => "ERROR", "message" => "Autorisation refusée"]);
         }
     }
 }
